+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 # 0.7.3
 
-- added option `csv-edit.autoColumnWithsIgnoreComments` to ignore comment cells for auto column sizing
+- added option `csv-edit.autoColumnWidthsIgnoreComments` to ignore comment cells for auto column sizing
 - `csv-edit.lastRowOrFirstRowNavigationBehavior` default is now `stop` (was `wrap`)
 - `csv-edit.lastColumnOrFirstColumnNavigationBehavior` default is now `stop` (was `wrap`)